+++ src/curvesimilarities/__init__.py
@@ -1,7 +1,7 @@
 """Curve similarity measures."""
 
 from .averagefrechet import afd, afd_owp, qafd, qafd_owp
-from .dtw import dtw, dtw_owp
+from .dtw import dtw, dtw_owp, sdtw, sdtw_owp
 from .frechet import dfd, fd
 from .integfrechet import ifd, ifd_owp
 
@@ -10,6 +10,8 @@ __all__ = [
     "dfd",
     "dtw",
     "dtw_owp",
+    "sdtw",
+    "sdtw_owp",
     "ifd",
     "ifd_owp",
     "afd",